feat!: ignore rollover evts during TDC2 fast mode
Rollover events are generally not useful for the end-user since their
main function is for readevents to perform fast epoch rollovers
correctly. fpfind supports ignoring these rollover events, and has been
added as a dependency, partly also because the timestamp parser there is
better maintained.

fpfind has a Python >=3.8 requirement, so this behaviour will only be
activated as such.