feat!: reintroduce Cython compilation
Cythonizing pulls compatible CPython headers for compilation, and doing
this prematurely is likely to cause issues when building wheels in
environments without said headers, e.g. when building 'manylinux' wheels
with 'cibuildwheel' across different Python versions. This is actually
addressed in numpy [1] and pandas [2], with the same concerns of CPython
version mismatch with the generated Cythonized files in sdist.

This reintroduces Cython compilation to fix this issue, at the expense
of increased compilation burden during manual installation (from
source), but now allows for wheels to be built cleanly.

[1]: https://github.com/numpy/numpy/pull/14453
[2]: https://github.com/pandas-dev/pandas/issues/28341