build!: migrate to pyproject.toml, drop Python 3.6
Library previously defines metadata in setup.py, which was appropriate
for setuptools pre-PEP518 when pyproject.toml was not specified. Modern
Python and tooling have since moved on to pyproject.toml, and is also
required for publishing metadata in PyPI.

A workaround involves duplicating metadata in both pyproject.toml and
setup.py, which is (1) error-prone, and (2) pointless since some
subpackages of S15lib actually no longer supports Python 3.6 due to the
use of Python 3.7+ functionality (e.g. dataclasses).

Section on which S15lib version to install has been added to assist old
users of Python 3.6.